Projeto_Furia/requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.1.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gevent==24.10.3
gevent-eventemitter==2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
groq==0.11.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.10.0
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==3.20.0
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
setuptools==75.2.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
st-theme==1.2.3
stack-data==0.6.3
steam==1.4.4
streamlit==1.37.0
streamlit-javascript==0.1.5
streamlit-navigation-bar==3.3.0
streamlit-option-menu==0.4.0
streamlit_carousel==1.1.2
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
ua-parser==0.18.0
urllib3==2.2.3
user-agents==2.2.0
vdf==3.4
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.1.1
